Dynamic battle state display.
Balance of powers in battle screen dynamically hides unneeded lines (works with instant result display to).
Instant loss display now in the same line as troop number.
Fleet and guard properly integrated in battle display.
Armies given to the battle function in the proper format: Interface of battle handler and battle calculator fixed.
Widened the battle screen and the boxes of the balance of power display.