Fixed event viability check.
Now events are properly marked as possible if the armie is located in such a way and has enough movement and height points to complete the move. This only checks one step (as that is how movement possibility is currently implemented). Other steps depending on another move event are marked as suspended.