Transport sinking implemented.
When transport ships are decimated transported troops are proportionally decimated so that they fit in their transports. Implemented an alternative damage function for foot and horse armies reducing troop, catapult, mount and officer counts proportionally so that a certain number of RP is removed. Changed the flooring functions in troop/officer/cata removing functions to round functions to better imitate the existing tools.