changes to new Gamestate lists in GodmodeFunction
in the Godmode changed the Productionbuilding and Riverlists to the new lists from the new Gamestate. also changed the buildings.ts to be able to change the type. for this i used the shortcuts for Getter and Setter that lets us use variables like normal but uses getter and setter in the background