Corrected the use of properties and constructor field declaration.
Refraining from using peroperties in order to obtain clear state of classes and fields when it comes to subclassing.
Properly defining the type field in the constructor in order to save two lines of code. Making type public (for now) so that godModeFunctions can still access it via the old interface provided by the properties.
Removing the leading underscores from names of constructor arguments because they are superfluous, don't match TypeScript convension and give me eye cancer.