Began refactoring the movement functions.
Moved the code from the move-function to the fleet and land army classes.
Heavily refactored the move functions to use higher order array functions.
Added a new class Move to later use in possible move prediction.
Added a new method to MobileEntity changePosition to set the old and new position to a certain value without using the usual move method (to be used for God-Mode).