Moved the possible move computation of rider armies to the LandArmy and Rider classes.
Set up the common component of rider and foot move computation in the method checkForPossibleMove of LandArmy.
Made an abstract method computeMoveCost in the LandArmy class to separate the move point cost computation (different for foot and riders) from the height cost computation, building fetching etc.
Moved the move point cost computation for riders to the method computeMoveCost of the RiderArmy class.