Started removing the uses of the old army object and using the new army classes.
Removed the inclusion of deprecated/army.js from the HTML file.
Removed the listOfArmies from phoenixclient.ts and changed all uses of it to the uses of GameState.armies.
Fixed errors caused by the change above in the phoenixclient.ts file (other files still to be fixed).
Army now keeps track of catapults shot.
MapEntity and subclasses now take precautions when passing their position (and old position) in and out of themselves so that nobody ends up with the same instance of a number tuple as they.