Fixed saving functions. But lot remains to be refactored.
Changed the type of changedBuildings in control variables.
The PhoenixEvent class now has an (abstract) method returning the content of the event in a JSON format. Needs to be implemented in every subclass.
Added new function to direction module converting a direction to a stirng.
Realms have a new function to convert their territory to an array of coordinates. Later used in transforming the list of realms to the borders JSON.