Refactored armyCoordinates.ts
Moved the clickedMoves funtion to the MobileEntity class. It now takes no arguments and is applied to the army it is the member of.
Created a new file multifieldFunctions in the gui folder and moven createMultifield and addToMultifield there.
Removed the now empty armyCoordinates file.
Fixed some imports to deal with refactored armyCoordinates.