Refactored armyInteraction.ts
The file armyInteraction.ts is now gone.
Shooting related functions are now in a new armies/shootingFunctions.ts file (namespace).
Some shooting functions are in the army class.
Some distance and obstruction recognition functions are now in HexFunctions.
Mount- and unMountWithParams functions are now called mount and dismount and are in FootArmy and RiderArmy.
Mount- and UnMountSelected functions are in the buttonFunctions.ts file.
Merging, splitting and transfering functions are also there.
DeleteArmy, generateArmyId and checkArmiesForLiveliness are rewamped and moved to a new libraries/armyFunctions.ts file.
DeleteArmy now takes an army and also deletes it form transporting fleets.
CheckArmiesForLiveliness now uses the deleteArmy function.