Refactored event status and event type mechanisms.
Added a new enumeration EventStatus to mark an event's status instead of a string.
Refactored event classes to use the new event status mechanism.
Removed the getType() function from events. This is not the proper way to check the type of an event. Use instanceof instead.
Renamed mergEvent.ts to mergeEvent.ts and adjusted files importing it.
Adjusted the visibility of makeEventListItem() to fit its use.
Moved the shootButtonLogic function out of the shootEvent class (what was it doing there in the first place?) and into buttonFunctions.