Implemented the board state viability check for MoveEvents.
Refactored checkEvent() function of MoveEvent to not use the unprocessedBattleAtContainingArmy function anymore.
Changed the signature of the checkForPossibleMove(direction) function to return a Move. Changed the clickedMoves function to add the returned moves to the list of possible moves.
Renamed the only list of events we will need in GameState to simply events.