Implemented the methods for converting event classes to JSON.
Reintroduced a getType function to the event class. It is used internally to get the string to send to server.
The getContent function is now protected and returns a string. It is used by the asJSON function to assemble the JSON representation of the event.
Introduced new asJSON function to event class assembling the JSON of the event.