Adjusted the saving functions to the new model.
All saving to server functions now send proper JSON. It is formatted with the redesigned DB server in mind (TODO: make appropriate changes on the DB server).
Added new constants to hold maximum build points for all buildings and maximum number of guards for walls.
Changed the visibility of of the typeAsString method of event classes.
Added a method returning the building in a JSON format to the building class.
Added getter and setter for various fields of building classes and generally fixed a lot of errors in class declarations.